Reorganize pdfs
