+++ src/pdfs/README.md
@@ -4,21 +4,31 @@ This directory contains all the blank PDFs that Namesake uses to fill forms, inc
 
 ## Naming and organization
 
-PDFs should be titled using all lowercase `kebab-case`. File names should begin with the form code, if one exists, followed by the full form title. For example, the form "Petition to Change Name of Adult" with the code "CJP 27" has the file name `cjp27-petition-to-change-name-of-adult.pdf`.
+Each PDF has its own folder under a jurisdiction (`ma`, `federal`, `ny`). The folder name is the PDF id in `kebab-case`. For example:
 
-If the form code contains any spaces or hyphens, those should be omitted. For example, the form code "CJ-D 400" is `cjd400`. Together with the title "Probate and Family Court Motion", the file name is `cjd400-probate-and-family-court-motion.pdf`.
+```
+src/pdfs/
+  ma/
+    affidavit-of-indigency/
+      index.ts           # typed PDF definition
+      schema.ts          # auto-generated form schema
+      affidavit-of-indigency.pdf
+      affidavit-of-indigency.test.ts
+```
 
-State-specific PDFs should be placed within a folder using the state's two-character abbreviation, like `/ma`. Federal forms, such as ones for Social Security or Passports, go in `/federal`.
+PDFs should be titled using all lowercase `kebab-case`. File names should begin with the form code, if one exists, followed by the full form title. For example, the form "Petition to Change Name of Adult" with the code "CJP 27" has the folder name `cjp27-petition-to-change-name-of-adult`.
+
+If the form code contains any spaces or hyphens, those should be omitted. For example, the form code "CJ-D 400" is `cjd400`. State-specific PDFs go in `/ma`, `/ny`, etc. Federal forms go in `/federal`.
 
 ## Defining PDF schemas
 
-Each `.pdf` file should be accompanied by a `.ts` definition of the same name containing a single `default export`. For example:
+Each PDF folder contains an `index.ts` with a single `default export`. For example:
 
 ```ts
-// cjp27-petition-to-change-name-of-adult.ts
+// index.ts
 import { definePdf } from "@/pdfs/utils/definePdf";
 import pdf from "./cjp27-petition-to-change-name-of-adult.pdf";
-import type { PdfFieldName } from "./cjp27-petition-to-change-name-of-adult.types";
+import type { PdfFieldName } from "./schema";
 
 export default definePdf<PdfFieldName>({
   id: "cjp27-petition-to-change-name-of-adult",
@@ -39,6 +49,8 @@ export default definePdf<PdfFieldName>({
 });
 ```
 
+Run `pnpm pdf:schema` to regenerate `schema.ts` from the PDF after adding or modifying form fields.
+
 ## Renaming PDF fields and adding new fields
 
 Sometimes original PDFs do not include form fields embedded within the PDF. Other PDFs do include form fields, but they are named poorly. In either case, you will want to make modifications to the PDF to add missing fields and name fields consistently.
@@ -47,39 +59,29 @@ To edit a PDF, download the original, then open it in the [Nutrient.io PDF Form 
 
 In general, it's good to name the fields so that they match the global constants for user form data. This will make mapping between the PDF field names and the names in our database much easier.
 
-Once you have finished editing the PDF, click the download button in the top right, rename the file, and add it to the appropriate directory within `/src/pdfs`.
+Once you have finished editing the PDF, click the download button in the top right, rename the file, and add it to the appropriate PDF folder within `/src/pdfs`.
 
 ## Testing PDFs
 
-Every PDF and definition should also include a test to validate that the form renders, checkboxes get checked, text fields get filled, etc.
+Each PDF folder includes a `[pdfid].test.ts` file to validate that the form renders, checkboxes get checked, text fields get filled, etc.
 
-Since we've already defined definitions for the data, this is pretty straightforward. Use `getPdfForm` to return a [PDFForm](https://pdf-lib.js.org/docs/api/classes/pdfform) object from `pdf-lib`, and then use methods like [getCheckBox](https://pdf-lib.js.org/docs/api/classes/pdfform#getcheckbox) and [getTextField](https://pdf-lib.js.org/docs/api/classes/pdfform#gettextfield) to test the values in the PDF.
-
-Take extra care to test any conditional logic around checkboxes.
+Use `getPdfForm` to return a [PDFForm](https://pdf-lib.js.org/docs/api/classes/pdfform) object from `pdf-lib`, then use methods like [getCheckBox](https://pdf-lib.js.org/docs/api/classes/pdfform#getcheckbox) and [getTextField](https://pdf-lib.js.org/docs/api/classes/pdfform#gettextfield) to test the values. Take extra care to test any conditional logic around checkboxes.
 
 ```ts
+// cjp27-petition-to-change-name-of-adult.test.ts
+import { getPdfForm } from "@/pdfs/utils/getPdfForm";
+import petitionToChangeNameOfAdult from ".";
+
 describe("CJP27 Petition to Change Name of Adult", () => {
-  const testData = {
-    newFirstName: "New",
-    newMiddleName: "Newly",
-    newLastName: "Name",
-    oldFirstName: "Old",
-    oldMiddleName: "Oldly",
-    oldLastName: "Name",
-    // More test data...
-  } as const;
+  const testData = { newFirstName: "New", oldFirstName: "Old", /* ... */ } as const;
 
   it("maps all fields correctly to the PDF", async () => {
     const form = await getPdfForm({
       pdf: petitionToChangeNameOfAdult,
       userData: testData,
     });
-
     expect(form.getTextField("newFirstName").getText()).toBe("New");
-    expect(form.getTextField("newMiddleName").getText()).toBe("Newly");
-    expect(form.getTextField("newLastName").getText()).toBe("Name");
-
-    // Test more fields...
-  })
+    // ...
+  });
 });
 ```
