+++ src/pdfs/federal/ss5-application-for-social-security-card/index.ts
@@ -2,7 +2,7 @@ import { definePdf } from "@/pdfs/utils/definePdf";
 import { formatBirthplaceCountryOrState } from "@/utils/formatBirthplaceCountryOrState";
 import { formatDateMMDDYYYY } from "@/utils/formatDateMMDDYYYY";
 import pdf from "./ss5-application-for-social-security-card.pdf";
-import type { PdfFieldName } from "./ss5-application-for-social-security-card.types";
+import type { PdfFieldName } from "./schema";
 
 export default definePdf<PdfFieldName>({
   id: "ss5-application-for-social-security-card",
