+++ src/pdfs/index.ts
@@ -2,7 +2,7 @@ import type { PDFDefinition, PDFId } from "@/constants/pdf";
 
 export async function getPdfDefinition(pdfId: PDFId) {
   const pdfModules = import.meta.glob(
-    ["/src/pdfs/!(utils)/**/!(*.test).ts", "!**/*.types.ts"],
+    ["/src/pdfs/!(utils)/**/index.ts"],
     { import: "default" },
   );
