+++ src/pdfs/ma/cjp27-petition-to-change-name-of-adult/cjp27-petition-to-change-name-of-adult.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { getPdfForm } from "@/pdfs/utils/getPdfForm";
-import petitionToChangeNameOfAdult from "../cjp27-petition-to-change-name-of-adult";
+import petitionToChangeNameOfAdult from ".";
 
 describe("CJP27 Petition to Change Name of Adult", () => {
   const testData = {
