+++ src/pdfs/ma/cjp34-cori-and-wms-release-request/index.ts
@@ -2,7 +2,7 @@ import { definePdf } from "@/pdfs/utils/definePdf";
 import { formatDateMMDDYYYY } from "@/utils/formatDateMMDDYYYY";
 import { joinNames } from "@/utils/joinNames";
 import pdf from "./cjp34-cori-and-wms-release-request.pdf";
-import type { PdfFieldName } from "./cjp34-cori-and-wms-release-request.types";
+import type { PdfFieldName } from "./schema";
 
 export default definePdf<PdfFieldName>({
   id: "cjp34-cori-and-wms-release-request",
