+++ src/pdfs/ma/cjp34-cori-and-wms-release-request/schema.ts
